control FindMove by passing string action
