--- chessVsSlimes/Assets/Scripts/Enemies/HungrySlime.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 17832325e93a64c07b98bc854a7c57bc
-timeCreated: 1540567708
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
